Rework class injection in session
*Breaking change*

This removes the injection of a request class into sessions and
replaces it with injecting a response class; getting custom response
classes was the original intent anyway (see the removed comment about
returning the result of calling `Request.get_response()`).
Additionally `Session`, `Request` and `CurlRequest` are now generic
over response type, to allow the exact response class to be understood
by type checkers.